fix cron scheduler bug & organize db models
