+++ .gitignore
@@ -18,10 +18,10 @@ todo.md
 private_key_dev.pem
 
 # Sqlite files in dev env
-database/kronus-dev.db
-database/kronus-dev.db-journal
-database/kronus-dev.db-shm
-database/kronus-dev.db-wal
+kronus-dev.db
+kronus-dev.db-journal
+kronus-dev.db-shm
+kronus-dev.db-wal
 
 # Dependency directories (remove the comment below to include it)
 # vendor/
