+++ models/base.go
@@ -1,0 +1,9 @@
+package models
+
+import "time"
+
+type BaseModel struct {
+	ID        uint `gorm:"primarykey"`
+	CreatedAt time.Time
+	UpdatedAt time.Time
+}
