+++ models/contact.go
@@ -1,0 +1,12 @@
+package models
+
+type Contact struct {
+	BaseModel
+	FirstName          string `json:"first_name"`
+	LastName           string `json:"last_name"`
+	PhoneNumber        string `json:"phone_number" validate:"required,e164" gorm:"not null;unique"`
+	Email              string `json:"email" validate:"required,email" gorm:"unique"`
+	IsEmergencyContact bool
+	UserID             uint             `gorm:"not null"`
+	EmergencyProbes    []EmergencyProbe `gorm:"constraint:OnUpdate:CASCADE,OnDelete:CASCADE;"`
+}
